Bootstrap idempotent

+++ bootstrap.sh
@@ -4,11 +4,16 @@ endpath="$HOME/.vimpeppers"
 
 echo "\n### Bootstrap started ###"
 
-echo "\n> clone repo\n"
-git clone http://github.com/gcapizzi/vimpeppers.git $endpath
+if [ -e $endpath/.git ]; then
+    echo "\n> update repo\n"
+    cd $endpath && git pull
+else
+    echo "\n> clone repo\n"
+    git clone http://github.com/gcapizzi/vimpeppers.git $endpath
+fi
 
 echo "> set up symlink\n"
-ln -s $endpath/vimrc $HOME/.vimrc
+ln -sf $endpath/vimrc $HOME/.vimrc
 
 if [ ! -e $HOME/.vim/bundle/vundle ]; then
     echo "> install Vundle\n"
